requirements.txt
flake8==3.7.5
flake8-blind-except==0.1.1
flake8-bugbear==18.8.0
flake8-builtins==1.4.1
flake8-commas==2.0.0
flake8-comprehensions==1.4.1
flake8-debugger==3.1.0
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2
flake8-print==3.1.0
flake8-quotes==1.0.0
flake8-string-format==0.2.3
flake8-todo==0.7
#flake8-annotations-coverage==0.0.2
#flake8-annotations-complexity==0.0.1
flake8-variables-names==0.0.1
dj-database-url==0.5.0
Django==2.1.7
django-configurations==2.1
django-debug-toolbar==1.11
django-extensions==2.1.5
factory-boy==2.11.1
Faker==1.0.2
psycopg2-binary==2.7.7
